Migrated sample app to KMP structure
