+++ lib/build.gradle.kts
@@ -3,7 +3,7 @@ import versioning.generateVersionName
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.compose)
+    alias(libs.plugins.compose.compiler)
 }
 
 val versionMajor = 4 // API Changes, adding big new feature, redesign the App
